Remove DataLoader in favor of s5tf-team/S5TF
