+++ Sources/Datasets/S5TFDatasetInfo.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public struct S5TFDatasetInfo {
+    var name: String
+    var version: String
+    var description: String
+    var homepage: URL
+}
